Slide out Burger menu for devices below 768px
Hid the burger menu using JavaScript and CSS styling to display none. and to make it appear used display flex. This is not ideal however I am still figuring out how to incorporate jQuery into my project. The animation when the side nav is hidden is choppy, but this is a quick fix to hide the hamburger when the menu is out.